backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.8.0
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0